take kdtree code to separate file
