+++ src/CMakeLists.txt
@@ -15,6 +15,8 @@ set(SOURCE_FILES
     "preview.cpp"
     "utilities.cpp"
     "utilities.h"
+    "kdtree.cpp"
+    "kdtree.h"
     )
 
 cuda_add_library(src
